moved files into working directory
team agreed to work from one folder
