+++ server.js
@@ -1,29 +1,39 @@
-//  require packages
-const path = require('path');
 const express = require('express');
 const session = require('express-session');
-const exphbs = require('express-handlebars');
-
-// connect routes
 const routes = require('./controllers');
+//Importing the handlebar package and helper codes
+const exphbs = require('express-handlebars');
 const helpers = require('./utils/helpers');
+const path = require('path');
 
 const sequelize = require('./config/connection');
+const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-// Set up Handlebars.js engine with custom helpers
+// Added 
 const hbs = exphbs.create({ helpers });
 
-app.use(session(sess));
+const sess = {
+  secret: 'Super secret secret',
+  cookie: {},
+  resave: false,
+  saveUninitialized: true,
+  store: new SequelizeStore({
+    db: sequelize
+  })
+};
 
-// Inform Express.js on which template engine to use
+app.use(session(sess));
+// Adding view engine
 app.engine('handlebars', hbs.engine);
+// Setting default view engine
 app.set('view engine', 'handlebars');
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+// Adding public folders to express static handlers
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(routes);
